requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-openai
pydantic
python-multipart
pillow
opencv-python
numpy